Add seperate service for handling invalidation
