+++ libs/api/domains/content-search/src/lib/enums/searchableContentTypes.ts
@@ -1,20 +1,5 @@
+import { SearchableContentTypes } from '@island.is/shared/types'
+export { SearchableContentTypes } from '@island.is/shared/types'
 import { registerEnumType } from '@nestjs/graphql'
-
-export enum SearchableContentTypes {
-  webArticle = 'webArticle',
-  webSubArticle = 'webSubArticle',
-  webLifeEventPage = 'webLifeEventPage',
-  webDigitalIcelandService = 'webDigitalIcelandService',
-  webDigitalIcelandCommunityPage = 'webDigitalIcelandCommunityPage',
-  webNews = 'webNews',
-  webAdgerdirPage = 'webAdgerdirPage',
-  webOrganizationSubpage = 'webOrganizationSubpage',
-  webOrganizationPage = 'webOrganizationPage',
-  webQNA = 'webQNA',
-  webLink = 'webLink',
-  webProjectPage = 'webProjectPage',
-  webManual = 'webManual',
-  webManualChapterItem = 'webManualChapterItem',
-}
 
 registerEnumType(SearchableContentTypes, { name: 'SearchableContentTypes' })
