+++ libs/content-search-indexer/src/lib/indexing.module.ts
@@ -5,11 +5,12 @@ import { CmsSyncModule } from '@island.is/cms'
 
 import { IndexingController } from './indexing.controller'
 import { IndexingService } from './indexing.service'
+import { CacheInvalidationService } from './cache-invalidation.service'
 
 @Module({
   // add the importer nestjs module to the imports array to have access to it's service in the indexing service
   imports: [CmsSyncModule],
   controllers: [IndexingController],
-  providers: [IndexingService, ElasticService],
+  providers: [CacheInvalidationService, IndexingService, ElasticService],
 })
 export class IndexingModule {}
