+++ libs/shared/types/src/index.ts
@@ -8,3 +8,4 @@ export * from './lib/organization'
 export * from './lib/static-text'
 export * from './lib/delegation'
 export * from './lib/environment'
+export * from './lib/searchable-content-types'
