+++ libs/shared/types/src/lib/searchable-content-types.ts
@@ -1,0 +1,16 @@
+export enum SearchableContentTypes {
+  webArticle = 'webArticle',
+  webSubArticle = 'webSubArticle',
+  webLifeEventPage = 'webLifeEventPage',
+  webDigitalIcelandService = 'webDigitalIcelandService',
+  webDigitalIcelandCommunityPage = 'webDigitalIcelandCommunityPage',
+  webNews = 'webNews',
+  webAdgerdirPage = 'webAdgerdirPage',
+  webOrganizationSubpage = 'webOrganizationSubpage',
+  webOrganizationPage = 'webOrganizationPage',
+  webQNA = 'webQNA',
+  webLink = 'webLink',
+  webProjectPage = 'webProjectPage',
+  webManual = 'webManual',
+  webManualChapterItem = 'webManualChapterItem',
+}
